+++ vitest.config.ts
@@ -3,6 +3,9 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     include: ['./test/tests/**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    chaiConfig: {
+      truncateThreshold: 100000,
+    },
     coverage: {
       statements: 85,
       branches: 83,